third task

+++ app/Http/Controllers/PostController.php
@@ -3,20 +3,25 @@
 namespace App\Http\Controllers;
 
 use App\Models\Post;
+use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\File;
 use Illuminate\Validation\Rules\Exists;
 use PHPUnit\Framework\Constraint\FileExists;
-
-use function Pest\Laravel\delete;
+use Laravel\Passport\Passport;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
+use Illuminate\Foundation\Bus\DispatchesJobs;
+use Illuminate\Foundation\Validation\ValidatesRequests;
+use Illuminate\Routing\Controller as BaseController;
 
 class PostController extends Controller
 {
+    use AuthorizesRequests, DispatchesJobs, ValidatesRequests;
     /**
      * Display a listing of the resource.
      */
     public function index()
-    {
+    {   
         $posts = Post::all();
         return view('posts.index', compact('posts'));
     }
@@ -25,7 +30,8 @@ class PostController extends Controller
      * Show the form for creating a new resource.
      */
     public function create()
-    {
+    {  
+        /* $this->authorize('manageUser',User::class); */
         return view('posts.create');
     }
 
@@ -34,6 +40,11 @@ class PostController extends Controller
      */
     public function store(Request $request)
     {
+        $request->validate([
+            'title' => 'string|required',
+            'description' => 'string|required',
+            'images' => 'array'
+        ]);
         $rightImages = [];
         if ($request->hasFile("images")) {
             foreach ($request->file("images") as $image) {
@@ -74,28 +85,21 @@ class PostController extends Controller
      * Update the specified resource in storage.
      */
     public function update(Request $request, Post $post)
-    {
+    {   
+        $request->validate([
+            'title' => 'string|required',
+            'description' => 'string|required',
+            'images' => 'array'
+        ]);
+        
         $imgsarray = json_decode($post->images);
-
         if ($request->hasFile('images')) {
             foreach ($imgsarray as $image) {
-               /*  $pathImage = public_path('/assets/images/' . $image);
-                if (File::exists($pathImage)) {
-                    File::delete($pathImage);
-                } */
-
                 $image_path = public_path("/assets/images/" . $image);
 
                 if (file_exists($image_path)) {
-
                     unlink($image_path);
                 }
-
-                /* if(file_exists($pathImage)){
-                    unlink($pathImage);
-                    delete($pathImage);
-                    
-                } */
             }
             $rightImages = [];
             foreach ($request->file('images')  as $image) {
@@ -120,6 +124,15 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        $this->authorize('manageUser', User::class); 
+        $imgsarray = json_decode($post->images);
+        foreach ($imgsarray as $image) {
+            $image_path = public_path("/assets/images/" . $image);
+
+            if (file_exists($image_path)) {
+                unlink($image_path);
+            }
+        } 
         $post->delete();
         return redirect()->route('posts.index');
     }
